requirements.txt
pandas==2.2.2
scikit-learn==1.5.1
numpy==1.26.4
pyyaml==6.0.2
joblib==1.4.2
scipy==1.13.1
xgboost==2.1.1
lightgbm==4.4.0
matplotlib==3.9.0
seaborn==0.13.2
tabulate==0.9.0
